Fix focus not always returning to expected element on active=false
This is a new fix for the issue identified in #54 after gaining
a better understanding of the issue.

The `returnFocusOnDeactivate` flag was being special-cased, but
not always, and the element to which focus should be returned,
which focus-trap-react attempts to manage on its own (see
class Constructor comment for reasoning), was not always correct.
For example, if the trap was deactivated, and then re-activated,
the 'last focused element before activation' wasn't updated at
the moment of re-activation, where it could've changed.