MVP support for React 19
focus-trap-react was already "ready" for React 19. All that was necessary
was to drop `prop-types`, which has been removed as a dependency.

It's also time to move forward and stop supporting ancient versions of
React. Therefore, the minimum supported version of React will now be
18.0.0, and I intend for that remain one version behind the latest
version going forward, provided the gap between the two is reasonably
surmountable.