Fix v11.0.0 types that removed the default export
Fixes #1396

My original intent was to remove the default export (a breaking change)
as part of the major 11.0.0 release but I totally flopped on that by
not also actually providing the named export in the code.

Now that the major release ship has sailed, this change fixes the types
by reinstating the default export in the typings but marking it
deprecated, while also providing the new named export alternative
as the way forward.